+++ doc/whats-new.rst
@@ -12,7 +12,8 @@ v2025.08.1 (unreleased)
 
 New Features
 ~~~~~~~~~~~~
-
+- Add convenience methods to :py:class:`~xarray.Coordinates` (:pull:`10318`)
+  By `Justus Magin <https://github.com/keewis>`_.
 - Added :py:func:`load_datatree` for loading ``DataTree`` objects into memory
   from disk. It has the same relationship to :py:func:`open_datatree`, as
   :py:func:`load_dataset` has to :py:func:`open_dataset`.